renderer: transpose RB_GetShaderTextureMatrix() before passing to GLSL.
Currently RB_GetShaderTextureMatrix() has not been modified to return
the matrix in column-major order due to it's use elsewhere in the
renderer (e.g. backEnd.lightTextureMatrix)

RB_GetShaderTextureMatrix() and RB_BakeTextureMatrixIntoTexgen() could
be optimized further to use column-major order everywhere if this proves
to be a significant cost during performance analysis.

Issue: #19
Issue: #30
Signed-off-by: Oliver McFadden <omcfadde@gmail.com>